Base,OpactOpt: move Amira readers to base module
