+++ modules/base/CMakeLists.txt
@@ -46,6 +46,8 @@ set(HEADER_FILES
     include/modules/base/datavisualizer/layertoimagevisualizer.h
     include/modules/base/datavisualizer/meshinformationvisualizer.h
     include/modules/base/datavisualizer/volumeinformationvisualizer.h
+    include/modules/base/io/amirameshreader.h
+    include/modules/base/io/amiravolumereader.h
     include/modules/base/io/binarystlwriter.h
     include/modules/base/io/datvolumesequencereader.h
     include/modules/base/io/datvolumewriter.h
@@ -199,6 +201,8 @@ set(SOURCE_FILES
     src/datavisualizer/layertoimagevisualizer.cpp
     src/datavisualizer/meshinformationvisualizer.cpp
     src/datavisualizer/volumeinformationvisualizer.cpp
+    src/io/amirameshreader.cpp
+    src/io/amiravolumereader.cpp
     src/io/binarystlwriter.cpp
     src/io/datvolumesequencereader.cpp
     src/io/datvolumewriter.cpp
