+++ modules/base/include/modules/base/io/amiravolumereader.h
@@ -1,9 +1,8 @@
 /*********************************************************************************
-/*********************************************************************************
  *
  * Inviwo - Interactive Visualization Workshop
  *
- * Copyright (c) 2024-2025 Inviwo Foundation
+ * Copyright (c) 2025 Inviwo Foundation
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without
@@ -29,32 +28,26 @@
  *********************************************************************************/
 #pragma once
 
-#include <modules/opactopt/opactoptmoduledefine.h>
+#include <modules/base/basemoduledefine.h>
 
-#include <inviwo/core/datastructures/geometry/mesh.h>
+#include <inviwo/core/datastructures/volume/volume.h>
 #include <inviwo/core/io/datareader.h>
-
-#include <memory>
-#include <filesystem>
 
 namespace inviwo {
 
 /**
  * \ingroup dataio
- * \brief AmiraMesh reader
+ * \brief AmiraMesh scalar field (volume) reader
  */
-class IVW_MODULE_OPACTOPT_API AmiraMeshReader : public DataReaderType<Mesh> {
+class IVW_MODULE_BASE_API AmiraVolumeReader : public DataReaderType<Volume> {
 public:
-    AmiraMeshReader();
-    AmiraMeshReader(const AmiraMeshReader& rhs) = default;
-    AmiraMeshReader& operator=(const AmiraMeshReader& that) = default;
-    virtual AmiraMeshReader* clone() const override;
-    virtual ~AmiraMeshReader() = default;
+    AmiraVolumeReader();
+    AmiraVolumeReader(const AmiraVolumeReader& rhs) = default;
+    AmiraVolumeReader& operator=(const AmiraVolumeReader& that) = default;
+    virtual AmiraVolumeReader* clone() const override;
+    virtual ~AmiraVolumeReader() = default;
 
-    virtual std::shared_ptr<Mesh> readData(const std::filesystem::path& filePath) override;
-
-private:
-    enum AmiraDataType { Lines, Vertices, VertexData };
+    virtual std::shared_ptr<Volume> readData(const std::filesystem::path& filePath) override;
 };
 
 }  // namespace inviwo
